app/requirements.txt
flask
flask_bootstrap
Flask-WTF
#bjoern
gunicorn
pyyaml
pytest
pytest-forked
pytest-cov
coverage
